:recycle: Make reusable all fields between send and request token
