+++ src/modules/SendToken/components/SelectApplicationsStep/styles.js
@@ -61,18 +61,22 @@ export default function getSendTokenSelectApplicationsStepStyles() {
   };
 }
 
-export const sendTokenRecipientAccountFieldStyles = {
-  containerStyle: {
-    paddingTop: 0,
-    paddingRight: 0,
-    paddingLeft: 0,
-    marginBottom: 16,
-    marginTop: 16,
-  },
-  inputLabel: {
-    marginBottom: 8
-  },
-  input: {
-    padding: 16
-  }
-};
+export function getSendTokenRecipientAccountFieldStyles(styles) {
+  return {
+    containerStyle: {
+      paddingTop: 0,
+      paddingRight: 0,
+      paddingLeft: 0,
+      marginBottom: 8,
+      ...styles?.container
+    },
+    inputLabel: {
+      marginBottom: 8,
+      ...styles?.inputLabel
+    },
+    input: {
+      padding: 16,
+      ...styles?.input
+    }
+  };
+}
