+++ src/navigation/index.js
@@ -27,7 +27,7 @@ import AddApplication from 'modules/BlockchainApplication/components/AddApplicat
 import AddApplicationSuccess from 'modules/BlockchainApplication/components/AddApplicationSuccess';
 import ApplicationDetail from 'modules/BlockchainApplication/components/ApplicationDetail';
 import SendToken from 'modules/SendToken';
-import Request from 'modules/Request';
+import RequestToken from 'modules/RequestToken';
 import TokensScreen from 'modules/Accounts/components/TokensScreen';
 import TransactionsHistory from 'modules/Transactions/components/TransactionsHistory';
 import navigationOptions from './navigationOptions';
@@ -165,7 +165,7 @@ const MainNavigator = () => {
           />
           <MainStack.Screen
             name="Request"
-            component={Request}
+            component={RequestToken}
             options={navigationOptions.NoHeader}
           />
           <MainStack.Screen
